Split api_routes.cpp into per-feature route modules
api_routes.cpp shrinks from 815 LOC of mixed routes to a 16-line aggregator
that delegates to:
  - features/meals/meals_routes — /api/meals, /api/ingredients, /api/plan
  - features/workouts/workouts_routes — /api/workouts, /api/workout-templates
  - integrations/google/google_routes — /auth/google/*, /, /api/calendar/*
  - core/http/static_routes — /<string>

New page = new register*Routes() in its feature folder + one line in
api_routes.cpp. setupRoutes() signature is unchanged so main.cpp doesn't
move. 101 tests still green.
