+++ src/core/http/api_routes.h
@@ -4,28 +4,21 @@
 
 #include <memory>
 
-#include "integrations/google/calendar_service.h"
 #include "core/config/config_parser.h"
 #include "core/db/db_manager.h"
-#include "integrations/google/google_oauth.h"
-#include "features/meals/meal_factory.h"
 #include "core/http/middleware.h"
+#include "features/meals/meal_factory.h"
+#include "integrations/google/calendar_service.h"
+#include "integrations/google/google_oauth.h"
 
 /**
- * @brief Configures all the REST API routes for the Meal Prep application.
+ * @brief Registers every HTTP route on the Crow app.
  *
- * This function registers the endpoints for viewing, adding, updating, and
- * deleting meals, as well as generating meal plans and serving static frontend
- * files.
- *
- * @param app The Crow web application instance.
- * @param dbManager Shared pointer to the database manager for database
- * operations.
- * @param factory Reference to the meal factory for instantiating meals.
- * @param config The application configuration containing settings such as email
- * credentials.
+ * Aggregates per-feature route modules (meals, workouts, Google integration,
+ * static files). New feature modules should add their own
+ * `register<Feature>Routes()` call here.
  */
-void setupRoutes(crow::App<RequestTimerMiddleware> &app, std::shared_ptr<DBManager> dbManager,
-                 MealFactory &factory, const Config &config,
-                 const std::shared_ptr<GoogleOAuth> &googleOAuth,
-                 const std::shared_ptr<CalendarService> &calendarService);
+void setupRoutes(crow::App<RequestTimerMiddleware>& app, std::shared_ptr<DBManager> dbManager,
+                 MealFactory& factory, const Config& config,
+                 const std::shared_ptr<GoogleOAuth>& googleOAuth,
+                 const std::shared_ptr<CalendarService>& calendarService);
