+++ src/core/http/static_routes.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <crow.h>
+
+#include "core/http/middleware.h"
+
+void registerStaticRoutes(crow::App<RequestTimerMiddleware>& app);
